+++ R/elr.R
@@ -1,24 +1,25 @@
 pos_res <- c("POSITIVE","REACTIVE","DETECTED","10828004","260373001","840533007","PCRP11","PDETD",
              "COVPRE","11214006","Positive for IgG","POS","DECTECTED","REA","PPOSI","REAC")
 
-#' ELR Terms - Positive
+#' ELR Terms
 #'
 #' A vector of SNOMED codes and keywords to be used with ELR recoding.
 #'
 #' @param collapse Boolean to collapse string separated by "|".
+#' @param add_in Vector of additional lab results.
 #'
-#' @rdname elr
 #' @return Vector of characters.
 #' @export
 #'
+#' @rdname elr
 #' @examples
-#' pos(collapse = FALSE)
+#' pos(collapse = TRUE)
 #' pos()
-pos <- function(collapse = FALSE){
+pos <- function(collapse = FALSE, add_in = NULL){
   if(collapse){
-    out <- paste0(pos_res, collapse = "|")
+    out <- paste0(c(pos_res, add_in), collapse = "|")
   } else {
-    out <- pos_res
+    out <- c(pos_res, add_in)
   }
   return(out)
 }
@@ -28,24 +29,17 @@ neg_res <- c("NEGATIVE","NON REACTIVE","NON-REACTIVE","NOT REACTIVE","NOT DETECT
              "NO DETECTED","NOT DETECTABLE","NOT DETECTE","NOT DETECTIVE","NOTDETECT","NOTDETECTED",
              "NR","NONREACTIVE","NOTREACTIVE")
 
-#' ELR Terms - Negative
-#'
-#' A vector of SNOMED codes and keywords to be used with ELR recoding.
-#'
-#' @param collapse Boolean to collapse string separated by "|".
-#'
-#' @rdname elr
-#' @return Vector of characters.
 #' @export
 #'
+#' @rdname elr
 #' @examples
-#' neg(collapse = FALSE)
+#' neg(collapse = TRUE)
 #' neg()
-neg <- function(collapse = FALSE){
+neg <- function(collapse = FALSE, add_in = NULL){
   if(collapse){
-    out <- paste0(neg_res, collapse = "|")
+    out <- paste0(c(neg_res, add_in), collapse = "|")
   } else {
-    out <- neg_res
+    out <- c(neg_res, add_in)
   }
   return(out)
 }