+++ R/wrap_labels.R
@@ -1,6 +1,9 @@
 #' Wrap Long Axis Labels
 #'
-#' Shorten total width of x-axis labels by breaking/wrapping text at forward slash or at "or".
+#' Break long axis labels at one or more "delimiters".
+#'
+#' @param x Input list of labels.
+#' @param delim Vector of characters or words to break label.
 #'
 #' @return Wrapped labels.
 #' @export
@@ -11,7 +14,11 @@
 #'
 #' ggplot(data = df, aes(x = group, y = score)) +
 #'  geom_col() +
-#'  scale_x_discrete(labels = wrap_labels())
-wrap_labels <- function() {
-  function(x){sub("\\/| or", "\\/\n", x)}
+#'  scale_x_discrete(labels = wrap_labels(delim = "or"))
+wrap_labels <- function(x, delim) {
+  if(length(delim)==1){
+    function(x){gsub(sprintf("(%s)", delim), "\\1\n", x)}
+  } else {
+    function(x){gsub(sprintf("(%s)", paste0(delim, collapse = "|")), "\\1\n", x)}
+  }
 }